+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.18.0] - 2023-05-05
+### Changed
+- Removed ingest checkpoints
+- Unique naming of temporary output directories
+
 ## [0.17.0] - 2022-07-22
 ### Added
 - Support for reading Apache Parquet files directly